add handle string and handle keepalive
